+++ TodoList/Commands/RemoveCommand.cs
@@ -1,29 +1,29 @@
 ﻿using System;
+using TodoList.Exceptions;
 
-namespace TodoList
+namespace TodoList;
+
+public class RemoveCommand : ICommand, IUndo
 {
-	public class RemoveCommand : ICommand, IUndo
+	public int Id { get; set; }
+	public string? TodosFilePath { get; set; }
+	private TodoItem? _removedItem;
+
+	public void Execute()
 	{
-		public int Id { get; set; }
-		public string? TodosFilePath { get; set; }
-		private TodoItem? _removedItem;
+		if (AppInfo.CurrentUserTodoList == null)
+			throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
 
-		public void Execute()
+		_removedItem = AppInfo.CurrentUserTodoList.GetById(Id);
+		AppInfo.CurrentUserTodoList.Remove(Id);
+	}
+
+	public void Unexecute()
+	{
+		if (_removedItem != null && AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
 		{
-			if (AppInfo.CurrentUserTodoList == null)
-				throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
-
-			_removedItem = AppInfo.CurrentUserTodoList.GetById(Id);
-			AppInfo.CurrentUserTodoList.Remove(Id); 
-		}
-
-		public void Unexecute()
-		{
-			if (_removedItem != null && AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
-			{
-				AppInfo.CurrentUserTodoList.AddExistingItem(_removedItem);
-				FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
-			}
+			AppInfo.CurrentUserTodoList.AddExistingItem(_removedItem);
+			FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
 		}
 	}
 }
