+++ TodoList/Commands/StatusCommand.cs
@@ -1,36 +1,36 @@
 ﻿using System;
+using TodoList.Exceptions;
 
-namespace TodoList
+namespace TodoList;
+
+public class StatusCommand : ICommand, IUndo
 {
-	public class StatusCommand : ICommand, IUndo
+	public int Id { get; set; }
+	public TodoStatus NewStatus { get; set; }
+	public string? TodosFilePath { get; set; }
+	private TodoStatus _previousStatus;
+
+	public void Execute()
 	{
-		public int Id { get; set; }
-		public TodoStatus NewStatus { get; set; }
-		public string? TodosFilePath { get; set; }
-		private TodoStatus _previousStatus;
+		if (AppInfo.CurrentUserTodoList == null)
+			throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
 
-		public void Execute()
+		var itemToUpdate = AppInfo.CurrentUserTodoList.GetById(Id);
+
+		if (itemToUpdate != null)
 		{
-			if (AppInfo.CurrentUserTodoList == null)
-				throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
-
-			var itemToUpdate = AppInfo.CurrentUserTodoList.GetById(Id);
-
-			if (itemToUpdate != null)
-			{
-				_previousStatus = itemToUpdate.Status;
-			}
-
-			AppInfo.CurrentUserTodoList.SetStatus(Id, NewStatus);
+			_previousStatus = itemToUpdate.Status;
 		}
 
-		public void Unexecute()
+		AppInfo.CurrentUserTodoList.SetStatus(Id, NewStatus);
+	}
+
+	public void Unexecute()
+	{
+		if (AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
 		{
-			if (AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
-			{
-				AppInfo.CurrentUserTodoList.SetStatus(Id, _previousStatus);
-				FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
-			}
+			AppInfo.CurrentUserTodoList.SetStatus(Id, _previousStatus);
+			FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
 		}
 	}
 }
