+++ TodoList/Commands/UpdateCommand.cs
@@ -1,36 +1,36 @@
 ﻿using System;
+using TodoList.Exceptions;
 
-namespace TodoList
+namespace TodoList;
+
+public class UpdateCommand : ICommand, IUndo
 {
-	public class UpdateCommand : ICommand, IUndo
+	public int Id { get; set; }
+	public string NewText { get; set; } = "";
+	public string? TodosFilePath { get; set; }
+	private string? _previousText;
+
+	public void Execute()
 	{
-		public int Id { get; set; }
-		public string NewText { get; set; } = "";
-		public string? TodosFilePath { get; set; }
-		private string? _previousText;
+		if (AppInfo.CurrentUserTodoList == null)
+			throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
 
-		public void Execute()
+		var itemToUpdate = AppInfo.CurrentUserTodoList.GetById(Id);
+
+		if (itemToUpdate != null)
 		{
-			if (AppInfo.CurrentUserTodoList == null)
-				throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
-
-			var itemToUpdate = AppInfo.CurrentUserTodoList.GetById(Id);
-
-			if (itemToUpdate != null)
-			{
-				_previousText = itemToUpdate.Text;
-			}
-
-			AppInfo.CurrentUserTodoList.Update(Id, NewText);
+			_previousText = itemToUpdate.Text;
 		}
 
-		public void Unexecute()
+		AppInfo.CurrentUserTodoList.Update(Id, NewText);
+	}
+
+	public void Unexecute()
+	{
+		if (_previousText != null && AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
 		{
-			if (_previousText != null && AppInfo.CurrentUserTodoList != null && TodosFilePath != null)
-			{
-				AppInfo.CurrentUserTodoList.Update(Id, _previousText);
-				FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
-			}
+			AppInfo.CurrentUserTodoList.Update(Id, _previousText);
+			FileManager.SaveTodos(AppInfo.CurrentUserTodoList, TodosFilePath);
 		}
 	}
 }
