+++ TodoList/Commands/ViewCommand.cs
@@ -1,20 +1,20 @@
 ﻿using System;
+using TodoList.Exceptions;
 
-namespace TodoList
+namespace TodoList;
+
+public class ViewCommand : ICommand
 {
-	public class ViewCommand : ICommand
+	public bool ShowIndex { get; set; }
+	public bool ShowStatus { get; set; }
+	public bool ShowUpdateDate { get; set; }
+	public bool ShowAll { get; set; }
+
+	public void Execute()
 	{
-		public bool ShowIndex { get; set; }
-		public bool ShowStatus { get; set; }
-		public bool ShowUpdateDate { get; set; }
-		public bool ShowAll { get; set; }
+		if (AppInfo.CurrentUserTodoList == null)
+			throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
 
-		public void Execute()
-		{
-			if (AppInfo.CurrentUserTodoList == null)
-				throw new AuthenticationException("Вы не авторизованы. Войдите в профиль, чтобы работать с задачами.");
-
-			AppInfo.CurrentUserTodoList.ViewCustom(ShowIndex, ShowStatus, ShowUpdateDate, ShowAll);
-		}
+		AppInfo.CurrentUserTodoList.ViewCustom(ShowIndex, ShowStatus, ShowUpdateDate, ShowAll);
 	}
 }
