+++ TodoList/Exceptions/AuthenticationException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class AuthenticationException : Exception
+	{
+		public AuthenticationException(string message) : base(message) { }
+	}
+}
