+++ TodoList/Exceptions/DuplicateLoginException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class DuplicateLoginException : Exception
+	{
+		public DuplicateLoginException(string message) : base(message) { }
+	}
+}
