+++ TodoList/Exceptions/InvalidArgumentException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class InvalidArgumentException : Exception
+	{
+		public InvalidArgumentException(string message) : base(message) { }
+	}
+}
