+++ TodoList/Exceptions/InvalidCommandException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class InvalidCommandException : Exception
+	{
+		public InvalidCommandException(string message) : base(message) { }
+	}
+}
