+++ TodoList/Exceptions/ProfileNotFoundException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class ProfileNotFoundException : Exception
+	{
+		public ProfileNotFoundException(string message) : base(message) { }
+	}
+}
