+++ TodoList/Exceptions/TaskNotFoundException.cs
@@ -1,0 +1,9 @@
+﻿using System;
+
+namespace TodoList.Exceptions
+{
+	public class TaskNotFoundException : Exception
+	{
+		public TaskNotFoundException(string message) : base(message) { }
+	}
+}
